requirements.txt
matplotlib~=3.6.0
numpy~=1.23.3
seaborn~=0.12.0
pandas~=1.5.0
scikit-learn~=1.1.1
Flask
flask-mysql
pyyaml
mysqlclient